refactor: move application setup logic from main.rs to startup.rs
