+++ rook/src/lib.rs
@@ -5,6 +5,7 @@ pub mod domain;
 pub mod email_client;
 pub mod git;
 pub mod link_checker;
+pub mod startup;
 
 pub use api::*;
 pub use configuration::*;
@@ -13,3 +14,4 @@ pub use domain::*;
 pub use email_client::*;
 pub use git::*;
 pub use link_checker::*;
+pub use startup::*;
